Time: 12 ms (39.13%) | Memory: 9.9 MB (91.90%) - LeetSync

+++ check-if-it-is-a-straight-line/check-if-it-is-a-straight-line.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool checkStraightLine(vector<vector<int>>& nums) {
-        double m;
+        float m;
         int n,x,y;
         n = nums.size();
         x = nums[0][0];
@@ -11,20 +11,16 @@ public:
         {
             for(int i=2;i<n;i++)
             {
-                if (x - nums[i][0] != 0)
-                {return 0;}
+                if (x - nums[i][0] != 0){return 0;}
             }
             return 1;
         }
         else
         {
-            m = (double(nums[1][1]-y))/denom;
+            m = (float(nums[1][1]-y))/denom;
             for(int i=2;i<n;i++)
             {
-                if (m != (double(nums[i][1]-y))/(nums[i][0]-x))
-                {
-                    return 0;
-                }
+                if (m != (double(nums[i][1]-y))/(nums[i][0]-x)){return 0;}
             }
             return 1;
         }
